refactor internal code structure
