+++ chimney/src/main/scala/io/scalaland/chimney/Transformer.scala
@@ -1,7 +1,7 @@
 package io.scalaland.chimney
 
-import io.scalaland.chimney.internal.ChimneyBlackboxMacros
 import io.scalaland.chimney.internal.dsl.TransformerDefinition
+import io.scalaland.chimney.internal.macros.ChimneyBlackboxMacros
 
 import scala.language.experimental.macros
