+++ chimney/src/main/scala/io/scalaland/chimney/internal/macros/PatcherMacros.scala
@@ -1,4 +1,7 @@
-package io.scalaland.chimney.internal
+package io.scalaland.chimney.internal.macros
+
+import io.scalaland.chimney.internal.utils.{DerivationGuards, EitherUtils, MacroUtils}
+import io.scalaland.chimney.internal.{DerivationConfig, DerivationError}
 
 import scala.reflect.macros.blackbox
