+++ chimney/src/main/scala/io/scalaland/chimney/internal/utils/DerivationGuards.scala
@@ -1,4 +1,4 @@
-package io.scalaland.chimney.internal
+package io.scalaland.chimney.internal.utils
 
 import scala.reflect.macros.blackbox
