+++ chimney/src/main/scala/io/scalaland/chimney/internal/utils/EitherUtils.scala
@@ -1,4 +1,4 @@
-package io.scalaland.chimney.internal
+package io.scalaland.chimney.internal.utils
 
 trait EitherUtils {
